Agregado de chequeo de outputs existentes, flags para salvar las metricas y flag verbose

+++ src/photoholmes/benchmark/model.py
@@ -19,20 +19,27 @@ class Benchmark:
     # Add documentation to class and methods
     def __init__(
         self,
-        save_output: bool = False,
+        save_output: bool = True,
+        save_metrics: bool = True,
         output_path: str = "output/",
         device: str = "cpu",
+        check_existing_output: bool = True,
+        verbose: bool = True,
     ):
-        self.save_output = save_output
+        self.save_output_flag = save_output
+        self.save_metrics_flag = save_metrics
         self.output_path = output_path
+        self.check_existing_output = check_existing_output
+        self.verbose = verbose
 
         if device.startswith("cuda") and not torch.cuda.is_available():
-            log.warning(
-                f"Requested device '{device}' is not available. Falling back to 'cpu'."
-            )
-            device = "cpu"
-        self.device = torch.device(device)
-        log.info(f"Using device: {self.device}")
+            if self.verbose:
+                log.warning(
+                    f"Requested device '{device}' is not available. Falling back to 'cpu'."
+                )
+        self.device = torch.device("cpu")
+        if self.verbose:
+            log.info(f"Using device: {self.device}")
 
         # TODO: set an attribute "output_keys" in the method class and use that
         # to determine whether to save the mask and heatmap or not
@@ -42,10 +49,11 @@ class Benchmark:
 
     def run(self, method: BaseMethod, dataset: BaseDataset, metrics: MetricCollection):
         if method.device != self.device:
-            log.warning(
-                f"Method device '{method.device}' does not match benchmark device '{self.device}'. "
-                f"Moving method to '{self.device}'"
-            )
+            if self.verbose:
+                log.warning(
+                    f"Method device '{method.device}' does not match benchmark device '{self.device}'. "
+                    f"Moving method to '{self.device}'"
+                )
             method.method_to_device(self.device)
 
         tampered_path = (
@@ -58,66 +66,80 @@ class Benchmark:
             dataset.__class__.__name__.lower(),
             tampered_path,
         )
+
         log.info("-" * 80)
         log.info("Running the benchmark")
         log.info("Benchmark configuration:")
         log.info(f"    Method: {method.__class__.__name__}")
         log.info(f"    Dataset: {dataset.__class__.__name__}")
-        log.info(f"    Metrics: {[metric for metric in metrics]}")
+        log.info("    Metrics:")
+        for metric in metrics:
+            log.info(f"       - {metric}")
         log.info(f"    Output path: {output_path}")
-        log.info(f"    Save output: {self.save_output}")
+        log.info(f"    Save output flag: {self.save_output_flag}")
+        log.info(f"    Save metrics flag: {self.save_metrics_flag}")
         log.info(f"    Device: {self.device}")
+        log.info(f"    Check existing output: {self.check_existing_output}")
+        log.info(f"    Verbose: {self.verbose}")
         log.info("-" * 80)
-
-        metrics_on_device = metrics.to("cpu", dtype=torch.float32)
-
-        heatmap_metrics = metrics_on_device.clone(prefix="heatmap")
-        mask_metrics = metrics_on_device.clone(prefix="mask")
-        detection_metrics = metrics_on_device.clone(prefix="detection")
+        if self.save_metrics_flag:
+            metrics_on_device = metrics.to("cpu", dtype=torch.float32)
+            heatmap_metrics = metrics_on_device.clone(prefix="heatmap")
+            mask_metrics = metrics_on_device.clone(prefix="mask")
+            detection_metrics = metrics_on_device.clone(prefix="detection")
 
         for data, mask, image_name in tqdm(dataset, desc="Processing Images"):
             # TODO: make a cleaner way to move the data to the device
             # (conditioned to the method or something)
-            data_on_device = self.move_to_device(data)
+            output = None
+            if self.check_existing_output:
+                output = self.check_for_existing_output(output_path, image_name)
 
-            mask = mask.to(self.device)
-            output = method.predict(**data_on_device)
+            if output is None:
+                data_on_device = self.move_to_device(data)
+                output = method.predict(**data_on_device)
+
             output = {
                 k: v.to("cpu") if isinstance(v, torch.Tensor) else v
                 for k, v in output.items()
             }
-            mask = mask.to("cpu")
+            if self.save_metrics_flag:
+                mask = mask.to("cpu")
+                if "detection" in output:
+                    detection_gt = (
+                        torch.tensor(int(torch.any(mask))).unsqueeze(0).to("cpu")
+                    )
+                    detection_metrics.update(output["detection"], detection_gt)
+                    self.save_detection = True
+                if "mask" in output:
+                    mask_metrics.update(output["mask"], mask)
+                    self.save_mask = True
+                if "heatmap" in output:
+                    heatmap_metrics.update(output["heatmap"], mask)
+                    self.save_heatmap = True
 
-            if "detection" in output:
-                detection_gt = torch.tensor(int(torch.any(mask))).unsqueeze(0).to("cpu")
-                detection_metrics.update(output["detection"], detection_gt)
-                self.save_detection = True
-            if "mask" in output:
-                mask_metrics.update(output["mask"], mask)
-                self.save_mask = True
-            if "heatmap" in output:
-                heatmap_metrics.update(output["heatmap"], mask)
-                self.save_heatmap = True
-
-            if self.save_output:
+            if self.save_output_flag:
                 self.save_pred_output(output_path, image_name, output)
 
         log.info("-" * 80)
-        if self.save_heatmap:
-            log.info("     - Saving heatmap metrics")
-            self.save_metrics(output_path, heatmap_metrics)
+        if self.save_metrics_flag:
+            if self.save_heatmap:
+                log.info("     - Saving heatmap metrics")
+                self.save_metrics(output_path, heatmap_metrics)
+            else:
+                log.info("     - No heatmap metrics to save")
+            if self.save_mask:
+                log.info("     - Saving mask metrics")
+                self.save_metrics(output_path, mask_metrics)
+            else:
+                log.info("     - No mask metrics to save")
+            if self.save_detection:
+                log.info("     - Saving detection metrics")
+                self.save_metrics(output_path, detection_metrics)
+            else:
+                log.info("     - No detection metrics to save")
         else:
-            log.info("     - No heatmap metrics to save")
-        if self.save_mask:
-            log.info("     - Saving mask metrics")
-            self.save_metrics(output_path, mask_metrics)
-        else:
-            log.info("     - No mask metrics to save")
-        if self.save_detection:
-            log.info("     - Saving detection metrics")
-            self.save_metrics(output_path, detection_metrics)
-        else:
-            log.info("     - No detection metrics to save")
+            log.info("     - Not saving metrics")
         log.info("-" * 80)
         log.info("Benchmark finished")
         log.info("-" * 80)
@@ -150,7 +172,8 @@ class Benchmark:
             ):
                 json_report[key] = [v.tolist() for v in value]
             else:
-                log.warning(f"Skipping metric '{key}' of type '{type(value)}'")
+                if self.verbose:
+                    log.warning(f"Skipping metric '{key}' of type '{type(value)}'")
 
         with open(
             os.path.join(metrics_path, f"{metrics.prefix}_report.json"), "w"
@@ -159,20 +182,65 @@ class Benchmark:
 
     def save_pred_output(self, output_path, image_name, output):
         image_save_path = os.path.join(output_path, "outputs", image_name)
-        os.makedirs(image_save_path, exist_ok=True)
+        # check if the files arrays and data.json already exist
+        if not self.check_existing_output or not os.path.exists(
+            os.path.join(image_save_path, "arrays.npz")
+        ):
+            os.makedirs(image_save_path, exist_ok=True)
 
-        array_like_dict = {}
-        non_array_like_dict = {}
+            array_like_dict = {}
+            non_array_like_dict = {}
 
-        for key, value in output.items():
-            if isinstance(value, (torch.Tensor)):
-                array_like_dict[key] = value.cpu()
-            elif isinstance(value, (list, np.ndarray)):
-                array_like_dict[key] = value
-            else:
-                non_array_like_dict[key] = value
+            for key, value in output.items():
+                if isinstance(value, (torch.Tensor)):
+                    array_like_dict[key] = value.cpu()
+                elif isinstance(value, (list, np.ndarray)):
+                    array_like_dict[key] = value
+                else:
+                    non_array_like_dict[key] = value
 
-        np.savez_compressed(os.path.join(image_save_path, "arrays"), **array_like_dict)
-        if non_array_like_dict:
-            with open(os.path.join(image_save_path, "data.json"), "w") as f:
-                json.dump(non_array_like_dict, f)
+            np.savez_compressed(
+                os.path.join(image_save_path, "arrays"), **array_like_dict
+            )
+            if non_array_like_dict:
+                with open(os.path.join(image_save_path, "data.json"), "w") as f:
+                    json.dump(non_array_like_dict, f)
+        elif self.verbose:
+            log.info(
+                f"Output for image '{image_name}' already exists. "
+                f"Skipping saving output."
+            )
+
+    def check_for_existing_output(self, output_path, image_name):
+        output_path = os.path.join(*output_path.split("/")[:-1])
+        folders = ["tampered_pristine/outputs", "tampered_only/outputs"]
+        for folder in folders:
+            if not os.path.exists(os.path.join(output_path, folder)):
+                continue
+            files = os.listdir(os.path.join(output_path, folder))
+            if image_name in files:
+                if self.verbose:
+                    log.info(
+                        f"Output for image '{image_name}' already exists. "
+                        f"Loading existing output."
+                    )
+                prior_output = np.load(
+                    os.path.join(output_path, folder, image_name, "arrays.npz"),
+                    allow_pickle=True,
+                )
+                prior_output = dict(prior_output)
+                prior_output = {
+                    k: torch.from_numpy(v) if isinstance(v, np.ndarray) else v
+                    for k, v in prior_output.items()
+                }
+                data_json_path = os.path.join(
+                    output_path, folder, image_name, "data.json"
+                )
+                if os.path.exists(data_json_path):
+                    with open(data_json_path, "r") as f:
+                        data_json = json.load(f)
+                    prior_output.update(data_json)
+                return prior_output
+        if self.verbose:
+            log.info(f"No prior output found for image '{image_name}'.")
+        return None
